requirements.txt
asgiref==3.7.2
attrs==23.1.0
botocore==1.31.54
Django==4.1.11
django-cors-headers==4.2.0
django-storages==1.14
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djongo==1.3.6
dnspython==2.4.2
drf-spectacular==0.25.1
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
Pillow==10.0.1
PyJWT==2.8.0
pymongo==3.13.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.10.2
s3transfer==0.6.2
six==1.16.0
sqlparse==0.2.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
gunicorn==21.2.0